Create Shortest Remaining Time First
This Algorithm is the preemptive version of shortest job first(SJF) scheduling. In  shortest remaining time first(SRTF), the execution of the process can be stopped after certain amount of time. At the arrival of every process, the short term scheduler schedules the process with the least remaining burst time among the list of available processes and the running process.